+++ test/index.js
@@ -1,20 +1,14 @@
 'use strict'
 
 const { TestDirector } = require('test-director')
-const checkGraphQLSchema = require('./checkGraphQLSchema.test')
-const checkGraphQLValidationRules = require('./checkGraphQLValidationRules.test')
-const checkOptions = require('./checkOptions.test')
-const errorHandler = require('./errorHandler.test')
-const execute = require('./execute.test')
-const isEnumerableObject = require('./isEnumerableObject.test')
 
 const tests = new TestDirector()
 
-isEnumerableObject(tests)
-checkOptions(tests)
-checkGraphQLValidationRules(tests)
-checkGraphQLSchema(tests)
-errorHandler(tests)
-execute(tests)
+require('./checkGraphQLSchema.test')(tests)
+require('./checkGraphQLValidationRules.test')(tests)
+require('./checkOptions.test')(tests)
+require('./errorHandler.test')(tests)
+require('./execute.test')(tests)
+require('./isEnumerableObject.test')(tests)
 
 tests.run()